Fixed warning from sklearn
